requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
sqlparse==0.5.1
psycopg2-binary==2.9.10
drf-spectacular==0.27.2
